Refactor services client
